requirements.txt
Babel==2.6.0
cliar==1.2.2
Django==2.1.7
django-appconf==1.0.3
django-compressor==2.2
django-countries==5.3.3
django-extra-views==0.11.0
django-haystack==2.8.1
django-oscar==1.6.7
django-phonenumber-field==2.0.1
django-tables2==1.21.2
django-treebeard==4.3
django-widget-tweaks==1.4.3
factory-boy==2.11.1
Faker==1.0.2
foliant==1.0.8
foliantcontrib.imagemagick==1.0.1
mock==2.0.0
pbr==5.1.3
phonenumbers==8.10.6
Pillow==5.4.1
prompt-toolkit==2.0.9
purl==1.4
pycountry==18.12.8
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
rcssmin==1.0.6
redis==3.2.1
rjsmin==1.0.12
six==1.12.0
sorl-thumbnail==12.3
text-unidecode==1.2
Unidecode==1.0.23
Wand==0.5.1
wcwidth==0.1.7
